+++ app/.dockerignore
@@ -1,0 +1,5 @@
+venv/
+__pycache__/
+*.pyc
+*.pyo
+.DS_Store
